fix: Integer 범위를 넘어가는 가격과 수량을 고려하여 타입 수정 #23

+++ src/main/java/com/onggiyonggi/domain/item/domain/Item.java
@@ -35,9 +35,9 @@ public class Item extends BaseEntity {
     private String name;
 
     @Column(nullable = false)
-    private Integer price;
+    private Long price;
 
     @Column(nullable = false)
-    private Integer count;
+    private Long count;
 
 }
